Working gnugol.cgi CGI on localhost. It's uglier than php, which I didn't think possible. I'd like to find a better cgi library for C.
The biggest issue that I'd like to tackle next is the
multiple listener problem for finding an appropriate
gnugol server. I grafted some multicast code in the
other day, what I'd like to find is a clean way to
have multiple listeners on various sockets, abstracted
out much like glib - but without the overhead of glib.