Chilling effects... no google v2 API.
Now that I have a clean way to implement more engines, I made the
mistake of thinking that I should add the latest google API.

Got the basic code for the new google API working...

Then read the license and tried to setup something that could use it.
There were too many checkboxes to click through.

Also realized that the new API was VERY SLOW - half the speed of the other one.

So, I added a disclaimer and I disabled the code by default.

I'm off to implement some engine with some license that doesn't
frighten me to death, now.