C callback functions and declarations are not my forte - busted code
Some... like murray gell-man - like to check in their code with all
the warts erased.

Others... like feynman and myself... like to make notes as to what
went wrong, and the solution, so they learn from mistakes in the future.

I've managed to NOT successfully declare a ptr to a function, somewhere,
in this commit. I'll fix it later, before I push.