Obsoleted *ad parameter, somewhat reluctantly.
While 99.99% of the time, *I* don't want ads, it would be good
to actually have a means of dealing with them.

However, it seems illogical that I'd have one per result, it's
more likely they would go in the header or footer, somehow, so
eliminating that param cleaned up the API.

Also fixed Makefile to only copy the shared objects to the sharedlibdir.