Reworked HTML entity conversion
Given that entities can also appear in the form '&#nnnn;' handle that case.
To streamline the code, convert the large mapping table to include the
Unicode code point instead of the UTF-8 string (which incidentally, matches
the original code the table was pulled from).

Also added a few calls to assert().  The code always passes in an entity
that starts with '&' so avoid the overhead of a check (when compiling with
NDEBUG).  I'd like to remove the check for the ending ';' as well, but since
the code that calls this needs to check for the trailing ';' but doesn't, so
keep the check in here for now (a missing trailing ';' makes for an invalid
HTML entity character).