commented out new TERM code
I think the idea is good - reformat the TERM mode output to fit
on a terminal screen. Being able to specify the max length of the
title and the snippet is also a good idea in the general case.

Terminal output CAN be beautiful, I'm sure.

The implementation, thus far, leaves much to be desired.