Fixed up the documentation a bit
Craig Guessford had contributed an awesome new logo as well as a
favicon, so most of the pages have those now. I'm not happy with the
default formatting I get from org at the moment, but...

Also moved a few pieces of the documentation from first person
to third, and other cleanups.