Non-crashing version of stackapps (for stack overflow.com)
I have a low barrier for checkins at this point. If it doesn't
crash, I'm going to push it out for people to try. It's
pretty much entirely independent of the rest of gnugol so it
shouldn't break anything.

stackapps is ORG MODE ONLY at this point, as the big thing I'm coping
with is that from a UI perspective:

url SHORT TITLE Description Snippet more closely matches the
kind of results I want to get from the web. Really long links
in org-mode are hard to look at.

This checkin also includes some extra gnugol.el functions that test
this engine out AND compare the results to just using google's site
specific search. It seems like what would work best would be a google
site search + the json output.